Utils refactoring
